// interfaces/userPortfolio.ts
// Generated by https://quicktype.io

export interface UserPortfolio {
    id:         number;
    name:       string;
    surname:    string;
    age:        number;
    city:       string;
    email:      string;
    title:      string;
    profile:    string;
    imageURL:   string;
    skills:     Skill[];
    formations: Formation[];
    projects:   Project[];
}

export interface Formation {
    id:              number;
    title:           string;
    educationEntity: string;
    inProgress:      boolean;
    startDate:       string;
    endDate:         string;
    certificateURL:  string;
}

export interface Project {
    id:            number;
    title:         string;
    description:   string;
    imageURL:      string;
    repositoryURL: string;
    deployURL:     string;
}

export interface Skill {
    id:        number;
    title:     string;
    typeSkill: TypeSkill;
}

export enum TypeSkill {
    Hard = "HARD",
    Soft = "SOFT",
}